Remove unnecessary usage of `unsafe` and replace some unsafe code with equivalent safe code.

The safe implementation of `iter::Bytes::slice_skip` proposed by this pull request has been benchmarked to be (slightly) slower than the currently used implementation, so if performance is an issue I can revert the changes to that particular method while still keeping my other changes.

`shrink` cannot be written without using unsafe (there are issues related to temporarily mutably borrowing the slice multiple times simultaneously), so I am ignoring it in this pull request. I might come back later and look at how `shrink` is used to see if something safer can be used instead.

I also ignored usage of `str::from_utf8_unchecked` since it seems apparent that whenever it is used in this crate the bytes passed in have already been checked for utf-8 compliance and so checking again would be redundant.