@seanmonstar 2021-03-15
Thanks so much for filing this! I'll take a look through now :)