@seanmonstar 2021-03-15
I was curious how the extra bounds checks would perform, so when I run the benchmarks of master vs this PR, here's what I see:

- **master**:
  
  ```
  test bench_httparse       ... bench:         569 ns/iter (+/- 103) = 1265 MB/s
  test bench_httparse_short ... bench:          67 ns/iter (+/- 1) = 1014 MB/s
  ```

- **PR**:
  
  ```
  test bench_httparse       ... bench:         701 ns/iter (+/- 67) = 1027 MB/s
  test bench_httparse_short ... bench:          80 ns/iter (+/- 1) = 850 MB/s
  ```